signer, accounts: move eip-712 types to apitypes package. fixes #23972
