+++ internal/ethapi/api.go
@@ -20,6 +20,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/ethereum/go-ethereum/signer/core/apitypes"
 	"math/big"
 	"strings"
 	"time"
@@ -1958,7 +1959,7 @@ func (api *PublicDebugAPI) TestSignCliqueBlock(ctx context.Context, address comm
 		return common.Address{}, err
 	}
 
-	signature, err := wallet.SignData(account, accounts.MimetypeClique, encoded)
+	signature, err := wallet.SignData(account, apitypes.MimetypeClique, encoded)
 	if err != nil {
 		return common.Address{}, err
 	}
